Moved CaseM cache updater into sepearte bean so it would not lock the cache
