+++ src/main/java/fi/otavanopisto/kuntaapi/server/integrations/casem/CaseMCache.java
@@ -2,15 +2,12 @@ package fi.otavanopisto.kuntaapi.server.integrations.casem;
 
 import java.util.ArrayList;
 import java.util.Collections;
-import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
 import java.util.concurrent.TimeUnit;
 import java.util.logging.Logger;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 
 import javax.annotation.PostConstruct;
 import javax.ejb.AccessTimeout;
@@ -19,11 +16,8 @@ import javax.enterprise.context.ApplicationScoped;
 import javax.inject.Inject;
 
 import org.apache.commons.lang3.StringUtils;
-import org.apache.commons.lang3.math.NumberUtils;
 
-import fi.otavanopisto.casem.client.ApiResponse;
 import fi.otavanopisto.casem.client.model.Node;
-import fi.otavanopisto.casem.client.model.NodeList;
 import fi.otavanopisto.casem.client.model.NodeName;
 import fi.otavanopisto.kuntaapi.server.controllers.IdentifierController;
 import fi.otavanopisto.kuntaapi.server.id.IdController;
@@ -47,9 +41,6 @@ public class CaseMCache {
   private Logger logger;
   
   @Inject
-  private CaseMApi caseMApi;
-  
-  @Inject
   private IdController idController;
   
   @Inject
@@ -66,11 +57,8 @@ public class CaseMCache {
   
   private Map<PageId, PageId> discoveredPageIds;
   
-  private boolean blooBloo;
-  
   @PostConstruct
   public void init() {
-    blooBloo = false;
     discoveredPageIds = new HashMap<>();
   }
 
@@ -103,6 +91,25 @@ public class CaseMCache {
     }
     
     return current;
+  }
+
+  public void cacheNode(OrganizationId organizationId, Long caseMRootNodeId, Node node) {
+    PageId pageId = caseMRootNodeId.equals(node.getNodeId()) ? null : translatePageId(toNodeId(node.getNodeId()), true);
+    PageId parentId = caseMRootNodeId.equals(node.getParentId()) ? null : translatePageId(toNodeId(node.getParentId()), true);
+    String pageCacheKey = getPageCacheKey(organizationId, pageId);
+    String parentCacheKey = getPageCacheKey(organizationId, parentId);
+    
+    List<PageId> cacheChildIds = treeCache.get(parentCacheKey);
+    if (cacheChildIds == null) {
+      cacheChildIds = new ArrayList<>(1);
+      cacheChildIds.add(pageId);
+      treeCache.put(parentCacheKey, cacheChildIds);
+    } else {
+      cacheChildIds.add(pageId);
+      treeCache.put(parentCacheKey, cacheChildIds);
+    }
+    
+    pageCache.put(pageCacheKey, translateNode(organizationId, caseMRootNodeId, node));
   }
   
   private Page findPageByParentAndSlug(OrganizationId organizationId, PageId parentId, String slug) {
@@ -209,31 +216,6 @@ public class CaseMCache {
     }
   }
   
-  public void refreshContents(OrganizationId organizationId) {
-    if (blooBloo) {
-      return;
-    }
-    
-    blooBloo = true;
-    
-    Long caseMRootNodeId = getCaseMRootNodeId(organizationId);
-    if (caseMRootNodeId == null) {
-      logger.severe(String.format("Organization %s CaseM root node is not defined", organizationId.toString()));
-    }
-    
-    List<Node> nodes = getChildNodes(organizationId, caseMRootNodeId, Collections.emptyList());
-    cacheNodeTree(organizationId, caseMRootNodeId, nodes, new ArrayList<>()); 
-  }
-  
-  private void cacheNodeTree(OrganizationId organizationId, Long caseMRootNodeId, List<Node> nodes, List<Long> caseMParentIds) {
-    for (Node node : nodes) {
-      cacheNode(organizationId, caseMRootNodeId, node);
-      List<Long> childCaseMParentIds = new ArrayList<>(caseMParentIds);
-      childCaseMParentIds.add(node.getNodeId());
-      cacheNodeTree(organizationId, caseMRootNodeId, getChildNodes(organizationId, caseMRootNodeId, childCaseMParentIds), childCaseMParentIds);
-    }
-  }
-  
   private PageId translatePageId(PageId pageId, boolean createMissing) {
     PageId result = discoveredPageIds.get(pageId);
     if (result != null) {
@@ -258,108 +240,4 @@ public class CaseMCache {
     return new PageId(CaseMConsts.IDENTIFIER_NAME, String.valueOf(nodeId));
   }
 
-  private void cacheNode(OrganizationId organizationId, Long caseMRootNodeId, Node node) {
-    PageId pageId = caseMRootNodeId.equals(node.getNodeId()) ? null : translatePageId(toNodeId(node.getNodeId()), true);
-    PageId parentId = caseMRootNodeId.equals(node.getParentId()) ? null : translatePageId(toNodeId(node.getParentId()), true);
-    String pageCacheKey = getPageCacheKey(organizationId, pageId);
-    String parentCacheKey = getPageCacheKey(organizationId, parentId);
-    
-    List<PageId> cacheChildIds = treeCache.get(parentCacheKey);
-    if (cacheChildIds == null) {
-      cacheChildIds = new ArrayList<>(1);
-      cacheChildIds.add(pageId);
-      treeCache.put(parentCacheKey, cacheChildIds);
-    } else {
-      cacheChildIds.add(pageId);
-      treeCache.put(parentCacheKey, cacheChildIds);
-    }
-    
-    pageCache.put(pageCacheKey, translateNode(organizationId, caseMRootNodeId, node));
-  }
-
-  private List<Node> getChildNodes(OrganizationId organizationId, Long caseMRootNodeId, List<Long> caseMParentIds) {
-    List<Node> result = new ArrayList<>();
-    Long skipToken = null;
-    
-    do {
-      NodeList nodeList = getChildNodeList(organizationId, caseMRootNodeId, caseMParentIds, skipToken);
-      if (nodeList == null) {
-        break;
-      }
-      
-      result.addAll(nodeList.getValue());
-      
-      skipToken = getSkipToken(nodeList.getOdataNextLink());
-    } while (skipToken != null);
-    
-    Collections.sort(result, new NodeComparator());
-    
-    return result;
-  }
-  
-  private NodeList getChildNodeList(OrganizationId organizationId, Long caseMRootNodeId, List<Long> caseMParentIds, Long skipToken) {
-    String pathQuery = getSubNodePath(caseMParentIds);
-    
-    ApiResponse<NodeList> response = caseMApi.getNodesApi(organizationId)
-      .listSubNodes(caseMRootNodeId, pathQuery, skipToken != null ? String.valueOf(skipToken) : null);
-    
-    if (!response.isOk()) {
-      logger.severe(String.format("Listing nodes by rootNode %d and pathQuery %s failed on [%d] %s", caseMRootNodeId, pathQuery, response.getStatus(), response.getMessage()));
-      return null;
-    } else {
-      return response.getResponse();
-    }
-  }
-  
-  private String getSubNodePath(List<Long> caseMParentIds) {
-    StringBuilder result = new StringBuilder();
-    
-    for (Long caseMParentId : caseMParentIds) {
-      result.append(String.format("SubNodes(%d)/", caseMParentId));
-    }
-    
-    result.append("SubNodes()");
-    
-    return result.toString();
-  }
-  
-  private Long getSkipToken(String nextLink) {
-    if (StringUtils.isBlank(nextLink)) {
-      return null;
-    }
-    
-    Pattern pattern = Pattern.compile("(.*\\$skiptoken=)([0-9]*)");
-    Matcher matcher = pattern.matcher(nextLink);
-    if (matcher.matches() && (matcher.groupCount() > 1)) {
-      return NumberUtils.createLong(matcher.group(2));
-    }
-    
-    return null;
-  }
-  
-  private Long getCaseMRootNodeId(OrganizationId organizationId) {
-    String rootNode = organizationSettingController.getSettingValue(organizationId, CaseMConsts.ORGANIZATION_SETTING_ROOT_NODE);
-    if (StringUtils.isNumeric(rootNode)) {
-      return NumberUtils.createLong(rootNode);
-    }
-    
-   return null;
-  }
-
-  private class NodeComparator implements Comparator<Node> {
-    @Override
-    public int compare(Node node1, Node node2) {
-      Integer sortOrder1 = node1.getSortOrder();
-      if (sortOrder1 == null) {
-        sortOrder1 = 0;
-      }
-      
-      Integer sortOrder2 = node1.getSortOrder();
-      if (sortOrder2 == null) {
-        sortOrder2 = 0;
-      }
-
-      return sortOrder1.compareTo(sortOrder2);
-    }
-  }
 }
